pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dsd-agent"
version = "0.1.0"
description = "Agent for populating Digital Solutioning Documents with architecture data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "python-pptx>=0.6.21",
    "anthropic>=0.39.0",
    "Pillow>=10.0.0",
]

[project.scripts]
dsd-agent = "dsd_agent.cli:main"

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
all = [
    "dsd-agent[google]",
]
